+++ application/src/main/java/ru/otus/basic/yampolskiy/servlets/ServletDispatcher.java
@@ -3,9 +3,9 @@ package ru.otus.basic.yampolskiy.servlets;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import ru.otus.basic.yampolskiy.servlets.exceptions.GlobalExceptionHandler;
-import ru.otus.basic.yampolskiy.servlets.filters.FilterChain;
-import ru.otus.basic.yampolskiy.servlets.filters.JwtAuthenticationFilter;
-import ru.otus.basic.yampolskiy.servlets.filters.PathAvailabilityFilter;
+import ru.otus.basic.yampolskiy.servlets.filtering.FilterChain;
+import ru.otus.basic.yampolskiy.servlets.filtering.JwtAuthenticationFilter;
+import ru.otus.basic.yampolskiy.servlets.filtering.PathAvailabilityFilter;
 import ru.otus.basic.yampolskiy.servlets.models.HttpServletRequest;
 import ru.otus.basic.yampolskiy.servlets.models.Route;
 import ru.otus.basic.yampolskiy.servlets.routing.ServletScanner;
